refactor(components): merge useSearchLocation \& useSelectSearchItem
МЕТА ДАННЫЕ: refactor 09
